# backend/internal/api/gen/user_to_group/create.go
// Code generated by apigen. DO NOT EDIT.

package user_to_group

import (
	"net/http"
	"time"

	"backend/db"
	"backend/internal/api/query"
	"backend/internal/api/rest"
	"backend/internal/api/validate"
	"backend/internal/authz"
	"backend/internal/syncer/types"

	"github.com/google/uuid"
	core "github.com/selectDb/dialect/core"

	syncgen "backend/internal/syncer/gen/user_to_group"
)

// createBodySchema is the request-body contract POST /user_to_groups validates
// against: the fields a client may set, their types and enums, and which are
// required. id is optional (a uuid when supplied; the handler generates one when
// omitted). Derived from the same field IR as the OpenAPI request schema. (A Go
// const can't hold a composite value, so this is a package var.)
var createBodySchema = validate.Schema{Fields: []validate.Field{
	{Name: "id", Kind: query.KindUUID},
	{Name: "user_id", Kind: query.KindUUID, Required: true},
	{Name: "group_id", Kind: query.KindUUID, Required: true},
}}

// Create handles POST /user_to_groups: create a row from the request body. The
// client may supply the id or let the server generate one. The write goes
// through the syncer's Apply, so tenancy, the cross-workspace FK guard, LWW, and
// audit emission match the sync path.
func Create() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		a := authz.ActorOf(r)
		if !rest.Gate(a, []string{core.ActionWorkspaceGroupsManage}) {
			rest.WriteError(w, http.StatusForbidden, "forbidden")
			return
		}
		body, ok := rest.DecodeBody(w, r)
		if !ok {
			return
		}
		if verr := validate.ForCreate(createBodySchema, body); verr != nil {
			rest.WriteValidationError(w, verr)
			return
		}
		id, _ := body["id"].(string)
		if id == "" { // server-assigned id when the client omits it
			id = uuid.NewString()
			body["id"] = id
		}
		if _, found, err := query.Get(r.Context(), db.GetDB(), resource, a.WorkspaceID, id); err != nil {
			rest.WriteError(w, http.StatusInternalServerError, "internal error")
			return
		} else if found {
			rest.WriteError(w, http.StatusConflict, singular+" already exists")
			return
		}
		c := types.Commit{
			ID:          id + ":create",
			CreatedAt:   time.Now(),
			Operation:   "create",
			TableName:   table,
			ObjectID:    id,
			Payload:     body,
			UserID:      a.UserID,
			WorkspaceID: a.WorkspaceID,
		}
		applied, _, err := syncgen.Apply(r.Context(), a.UserID, c, time.Time{})
		if err != nil {
			rest.WriteWriteError(w, singular, err)
			return
		}
		if !applied { // LWW rejected it (e.g. a soft-deleted row holds this id)
			rest.WriteError(w, http.StatusConflict, singular+" already exists")
			return
		}
		row, found, err := query.Get(r.Context(), db.GetDB(), resource, a.WorkspaceID, id)
		if err != nil {
			rest.WriteError(w, http.StatusInternalServerError, "internal error")
			return
		}
		if !found {
			w.WriteHeader(http.StatusCreated)
			return
		}
		rest.WriteJSON(w, http.StatusCreated, row)
	}
}